+++ src/love2d/mouse.cpp
@@ -1,12 +1,12 @@
 #include <all.hpp>
 
-int kame::love2d::mouse::getX()
+float kame::love2d::mouse::getX()
 {
     assert(kame::love2d::detail::Context::getInstance().isValid());
     return kame::love2d::detail::Context::getInstance().win->getState().mouseX;
 }
 
-int kame::love2d::mouse::getY()
+float kame::love2d::mouse::getY()
 {
     assert(kame::love2d::detail::Context::getInstance().isValid());
     return kame::love2d::detail::Context::getInstance().win->getState().mouseY;
